+++ src/Yaat.Sim/CommandQueue.cs
@@ -13,6 +13,7 @@ public enum BlockTriggerType
     DistanceFinal,
     OnHandoff,
     AtGroundEntity,
+    EnteringHoldingAfterExit,
 }
 
 public class BlockTrigger